Структурирован код и сам проект с учётом замечаний после ревью.
