+++ ImageFeed/Extensions/URLRequest+Extensions.swift
@@ -1,0 +1,20 @@
+//
+//  URLRequest+Extensions.swift
+//  ImageFeed
+//
+//  Created by Chingiz on 04.10.2023.
+//
+
+import Foundation
+
+extension URLRequest {
+    static func makeHTTPRequest(
+        path: String,
+        httpMethod: String,
+        baseURL: URL = DefaultBaseURL!
+    ) -> URLRequest {
+        var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
+        request.httpMethod = httpMethod
+        return request
+    }
+}
