+++ ImageFeed/Networking/NetworkError.swift
@@ -1,0 +1,14 @@
+//
+//  NetworkError.swift
+//  ImageFeed
+//
+//  Created by Chingiz on 04.10.2023.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case httpStatusCode(Int)
+    case urlRequestError(Error)
+    case urlSessionError
+}
