+++ ImageFeed/Networking/OAuth2TokenStorage.swift
@@ -7,7 +7,7 @@
 
 import Foundation
 
-class OAuth2TokenStorage {
+final class OAuth2TokenStorage {
     private let tokenKey = "BearerToken"
     
     var token: String? {
